totp : move TOTP-MCU to TOTP, update README and example
